+++ jekyll/assets/js/main.js
@@ -118,4 +118,63 @@ document.addEventListener("DOMContentLoaded", function () {
             alert(`REMOVED = Key: '${e.detail.key}' | Text: '${e.detail.value}'`);
         });
     }
+
+    if (document.querySelector("#option-animation")) {
+        let select = bootstrapNiceSelect.BootstrapNiceSelect("#option-animation");
+        document.querySelector("#option-animation-hide").addEventListener("click", function () {
+            select.bootstrapNiceSelect("hide");
+        });
+        document.querySelector("#option-animation-show").addEventListener("click", function () {
+            select.bootstrapNiceSelect("show");
+        });
+    }
+
+    if (document.querySelector("#methods-animation")) {
+        let select = bootstrapNiceSelect.BootstrapNiceSelect("#methods-animation");
+        document.querySelector("#methods-animation-hide").addEventListener("click", function () {
+            select.bootstrapNiceSelect("hide");
+        });
+        document.querySelector("#methods-animation-show").addEventListener("click", function () {
+            select.bootstrapNiceSelect("show");
+        });
+    }
+
+    if (document.querySelector("#methods-destroy")) {
+        let select = bootstrapNiceSelect.BootstrapNiceSelect("#methods-destroy");
+        document.querySelector("#methods-destroy-destroy").addEventListener("click", function () {
+            select.bootstrapNiceSelect("destroy");
+        });
+        document.querySelector("#methods-destroy-recreate").addEventListener("click", function () {
+            select = bootstrapNiceSelect.BootstrapNiceSelect("#methods-destroy");
+        });
+    }
+
+    if (document.querySelector("#methods-overlay")) {
+        let select = bootstrapNiceSelect.BootstrapNiceSelect("#methods-overlay");
+        document.querySelector("#methods-overlay-open").addEventListener("click", function () {
+            select.bootstrapNiceSelect("open", "A");
+        });
+        document.querySelector("#methods-overlay-close").addEventListener("click", function () {
+            select.bootstrapNiceSelect("close");
+        });
+    }
+
+    if (document.querySelector("#methods-select")) {
+        let select = bootstrapNiceSelect.BootstrapNiceSelect("#methods-select");
+        document.querySelector("#methods-select-select").addEventListener("click", function () {
+            select.bootstrapNiceSelect("select", { id: "Apple", text: "Apple" }, { id: "Lemon", text: "Lemon" }, { id: "Garlic", text: "Garlic" });
+        });
+
+        document.querySelector("#methods-select-select-all").addEventListener("click", function () {
+            select.bootstrapNiceSelect("selectAll");
+        });
+
+        document.querySelector("#methods-select-deselect").addEventListener("click", function () {
+            select.bootstrapNiceSelect("deselect", "Lemon", "Cabbages");
+        });
+
+        document.querySelector("#methods-select-deselect-all").addEventListener("click", function () {
+            select.bootstrapNiceSelect("deselectAll");
+        });
+    }
 });